+++ R/expr-list.R
@@ -193,11 +193,11 @@ expr_list_get <- function(index, ..., null_on_oob = TRUE) {
 #' index, use [`$list$get()`][expr_list_get]. The indices may be defined in a
 #' single column, or by sub-lists in another column of dtype List.
 #'
-#' @param indices An Expr or something coercible to an Expr, that can return
-#'   several indices. Values are 0-indexed (so index 0 would return the
-#'   first item of every sub-list) and negative values start from the end (index
-#'   `-1` returns the last item). If the index is out of bounds, it will return
-#'   a `null`. Strings are parsed as column names.
+#' @param indices An Expr or something coercible to an Expr of datatype List,
+#' (see examples). Values are 0-indexed (so index 0 would return the first item
+#' of every sub-list) and negative values start from the end (index `-1` returns
+#' the last item). If the index is out of bounds, it will return a `null`.
+#' Strings are parsed as column names.
 #' @inheritParams expr_list_get
 #' @inheritParams rlang::args_dots_empty
 #' @inherit as_polars_expr return
@@ -213,6 +213,11 @@ expr_list_get <- function(index, ..., null_on_oob = TRUE) {
 #'
 #' df$with_columns(
 #'   gathered = pl$col("a")$list$gather(list(2L), null_on_oob = TRUE)
+#' )
+#'
+#' # To select different indices per row:
+#' df$with_columns(
+#'   gathered = pl$col("a")$list$gather(list(2L, c(0L, 3L), 1L), null_on_oob = TRUE)
 #' )
 #'
 #' # Indices must be an List(Int/Uint) type to work.
@@ -514,8 +519,10 @@ expr_list_tail <- function(n = 5L) {
 #'   pl$col("b")$list$eval(pl$element()$str$join(" "))$list$first()
 #' )
 expr_list_eval <- function(expr) {
-  self$`_rexpr`$list_eval(as_polars_expr(expr)$`_rexpr`) |>
-    wrap()
+  wrap({
+    check_polars_expr(expr)
+    self$`_rexpr`$list_eval(expr$`_rexpr`)
+  })
 }
 
 #' Evaluate whether all boolean values in a sub-list are true
@@ -875,4 +882,36 @@ expr_list_drop_nulls <- function() {
 expr_list_count_matches <- function(element) {
   self$`_rexpr`$list_count_matches(as_polars_expr(element, as_lit = TRUE)$`_rexpr`) |>
     wrap()
+}
+
+#' Run any polars aggregation expression against the lists' elements
+#'
+#' This looks similar to [`$list$eval()`][expr_list_eval], but the key
+#' difference is that `$list$agg()` automatically explodes the list if the
+#' expression inside returns a scalar (while `$list$eval()` always returns a
+#' list).
+#'
+#' @inheritParams expr_list_eval
+#'
+#' @inherit as_polars_expr return
+#' @examples
+#' df <- pl$DataFrame(a = list(c(1, NA), c(42, 13), c(NA, NA)))
+#'
+#' # The column "null_count" has dtype u32 because `$null_count()` returns a
+#' # scalar for each sub-list. Using `$list$eval()` instead would return a
+#' # column with dtype list(u32).
+#' df$with_columns(
+#'   null_count = pl$col("a")$list$agg(pl$element()$null_count())
+#' )
+#'
+#' # The column "no_nulls" has dtype list(u32) because the expression doesn't
+#' # guarantee to return a scalar.
+#' df$with_columns(
+#'   no_nulls = pl$col("a")$list$agg(pl$element()$drop_nulls())
+#' )
+expr_list_agg <- function(expr) {
+  wrap({
+    check_polars_expr(expr)
+    self$`_rexpr`$list_agg(expr$`_rexpr`)
+  })
 }